pyproject.toml
[build-system]
requires = ["hatchling>=1.25", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "depth-anything-3"
version = "0.0.0"
description = "Depth Anything 3"
readme = "README.md"
requires-python = ">=3.10, <3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Your Name" }]

dependencies = [
    "pre-commit",
    "trimesh",
    "torch>=2",
    "torchvision",
    "einops",
    "huggingface_hub",
    "imageio",
    "imageio-ffmpeg",
    "numpy<2",
    "opencv-python",
    "xformers",
    "open3d",
    "fastapi",
    "uvicorn",
    "requests",
    "pillow",
    "omegaconf",
    "evo",
    "e3nn",
    "nerfstudio>=1.0.0; sys_platform == 'linux'",
    "plyfile",
    "pillow_heif",
    "safetensors",
    "uvicorn",
    "typer>=0.9.0",
    "moviepy==1.0.3",
    "gsplat @ git+https://github.com/nerfstudio-project/gsplat.git@v0.1.11 ; sys_platform == 'linux'",
    "pycolmap",
]

[project.optional-dependencies]
app = [
    "gradio>=5; python_version >= '3.10'",
    "pillow>=9.0; python_version >= '3.10'",
] # requires python>=3.10
streaming = [
    "faiss-gpu; sys_platform == 'linux'",
    "faiss-cpu; sys_platform == 'darwin'",
    "pandas",
    "prettytable",
    "numba",
    "pypose",
]
all = ["depth-anything-3[app,streaming]"]


[project.scripts]
da3 = "depth_anything_3.cli:app"

[project.urls]
Homepage = "https://github.com/ByteDance-Seed/Depth-Anything-3"

[tool.uv.extra-build-dependencies]
gsplat = ["torch"]

[tool.uv]
environments = [
  "sys_platform == 'linux' and platform_machine == 'x86_64' and python_full_version < '3.11'",
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/depth_anything_3"]

[tool.hatch.build.targets.sdist]
include = [
  "/README.md",
  "/pyproject.toml",
  "/src/depth_anything_3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.mypy]
plugins = ["jaxtyping.mypy_plugin"]

[tool.black]
line-length = 99
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["bson","cruise","cv2","dataloader","diffusers","omegaconf","tensorflow","torch","torchvision","transformers","gsplat"]
known_first_party = ["common", "data", "models", "projects"]
sections = ["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
skip_gitignore = true
line_length = 99
no_lines_before="THIRDPARTY"
